Fix Problem.cpp: PAM->Lloyd rename, size_t audit, triangular iteration, static vars, throw fix
- Rename cluster_by_kMedoidsPAM -> cluster_by_kMedoidsLloyd (scientifically accurate)
- Data::size() returns size_t instead of int (prevents overflow at N>2.1B)
- fillDistanceMatrix uses triangular iteration with O(1) index decode
- calculateMedoids uses local vectors instead of static (thread safety)
- Problem_IO.cpp: throw std::runtime_error instead of throw 1
- Updated all callers for size_t compatibility

+++ dtwc/Data.hpp
@@ -32,9 +32,9 @@ struct Data
 
   /**
    * @brief Returns the number of data points.
-   * @return Integer representing the size of the data vector.
+   * @return The size of the data vector (size_t).
    */
-  auto size() const { return static_cast<int>(p_vec.size()); }
+  auto size() const { return p_vec.size(); }
 
   Data() = default; //!< Default constructor
